+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [[v0.0.3]](https://github.com/mlange-42/arche-pixel/compare/v0.0.2...v0.0.3)
+
+### Breaking changes
+
+* Renamed `Window.Add` to `Window.AddDrawer` (#8)
+
+### Features
+
+* Adds `Image` plot for plotting grads and matrices (#8)
+* Additional drawers can be added to all plots, and draw over the plot's drawing (#8)
+
 ## [[v0.0.2]](https://github.com/mlange-42/arche-pixel/compare/v0.0.1...v0.0.2)
 
 ### Other